Remove redundant api endpoint
